+++ gameServer/common/mj/mj_base/interface.go
@@ -50,7 +50,7 @@ type DataManager interface {
 	GetUserCardIndex(ChairId int) []int
 	GetCurrentUser() int //当前出牌用户
 	GetRoomId() int
-	GetCreater() int
+	GetCreater() int64
 	GetProvideUser() int
 	IsActionDone() bool
